requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.2
